+++ ewm/src/main/java/ru/practicum/ewm/event/dto/EventShortDto.java
@@ -21,7 +21,7 @@ public class EventShortDto {
 
     private CategoryDto category;
 
-    private Integer confirmedRequests;
+    private long confirmedRequests;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
@@ -32,5 +32,5 @@ public class EventShortDto {
 
     private String title;
 
-    private Integer views;
+    private long views;
 }
